Unify ChessGame component with FEN/PGN controls
